+++ src/main/java/carpet/commands/DrawCommand.java
@@ -1,7 +1,7 @@
 package carpet.commands;
 
 import carpet.CarpetSettings;
-import carpet.settings.SettingsManager;
+import carpet.utils.CommandHelper;
 import carpet.utils.Messenger;
 import com.google.common.collect.Lists;
 import com.mojang.brigadier.CommandDispatcher;
@@ -37,7 +37,7 @@ public class DrawCommand
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher)
     {
         LiteralArgumentBuilder<CommandSourceStack> command = literal("draw").
-                requires((player) -> SettingsManager.canUseCommand(player, CarpetSettings.commandDraw)).
+                requires((player) -> CommandHelper.canUseCommand(player, CarpetSettings.commandDraw)).
                 then(literal("sphere").
                         then(argument("center", BlockPosArgument.blockPos()).
                                 then(argument("radius", IntegerArgumentType.integer(1)).
