+++ src/main/java/carpet/commands/ProfileCommand.java
@@ -1,7 +1,8 @@
 package carpet.commands;
 
 import carpet.CarpetSettings;
-import carpet.settings.SettingsManager;
+import carpet.utils.CommandHelper;
+
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.builder.LiteralArgumentBuilder;
 import net.minecraft.commands.CommandSourceStack;
@@ -18,7 +19,7 @@ public class ProfileCommand
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher)
     {
         LiteralArgumentBuilder<CommandSourceStack> literalargumentbuilder = literal("profile").
-                requires((player) -> SettingsManager.canUseCommand(player, CarpetSettings.commandProfile)).
+                requires((player) -> CommandHelper.canUseCommand(player, CarpetSettings.commandProfile)).
                 executes( (c) -> healthReport(c.getSource(), 100)).
                 then(literal("health").
                         executes( (c) -> healthReport(c.getSource(), 100)).
