+++ src/main/java/carpet/settings/RuleCategory.java
@@ -1,22 +1,59 @@
 package carpet.settings;
 
+/**
+ * @deprecated Use {@link carpet.api.settings.RuleCategory} instead. Should be as simple as changing to that in the imports
+ *
+ */
+@Deprecated(forRemoval = true)
 public class RuleCategory
 {
-    public static final String BUGFIX = "bugfix";
-    public static final String SURVIVAL = "survival";
-    public static final String CREATIVE = "creative";
-    public static final String EXPERIMENTAL = "experimental";
-    public static final String OPTIMIZATION = "optimization";
-    public static final String FEATURE = "feature";
-    public static final String COMMAND = "command";
-    public static final String TNT = "tnt";
-    public static final String DISPENSER = "dispenser";
-    public static final String SCARPET = "scarpet";
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#BUGFIX} instead
+     */
+    public static final String BUGFIX = carpet.api.settings.RuleCategory.BUGFIX;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#SURVIVAL} instead
+     */
+    public static final String SURVIVAL = carpet.api.settings.RuleCategory.SURVIVAL;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#CREATIVE} instead
+     */
+    public static final String CREATIVE = carpet.api.settings.RuleCategory.CREATIVE;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#EXPERIMENTAL} instead
+     */
+    public static final String EXPERIMENTAL = carpet.api.settings.RuleCategory.EXPERIMENTAL;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#OPTIMIZATION} instead
+     */
+    public static final String OPTIMIZATION = carpet.api.settings.RuleCategory.OPTIMIZATION;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#FEATURE} instead
+     */
+    public static final String FEATURE = carpet.api.settings.RuleCategory.FEATURE;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#COMMAND} instead
+     */
+    public static final String COMMAND = carpet.api.settings.RuleCategory.COMMAND;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#TNT} instead
+     */
+    public static final String TNT = carpet.api.settings.RuleCategory.TNT;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#DISPENSER} instead
+     */
+    public static final String DISPENSER = carpet.api.settings.RuleCategory.DISPENSER;
+    /**
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#SCARPET} instead
+     */
+    public static final String SCARPET = carpet.api.settings.RuleCategory.SCARPET;
     /**
      * Rules with this {@link RuleCategory} will have a client-side
      * counterpart, so they can be set independently without the server
      * being Carpet's
+     * @deprecated Use {@link carpet.api.settings.RuleCategory#CLIENT} instead
      */
-    public static final String CLIENT = "client";
+    @Deprecated(forRemoval = true)
+    public static final String CLIENT = carpet.api.settings.RuleCategory.CLIENT;
 
 }
